new version

+++ src/main/java/com/xiia/indicatori/repositories/AccountsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.xiia.indicatori.domain.Account;
 
-public interface AccountsRepository extends JpaRepository<Account, Long> {
+public interface AccountsRepository extends JpaRepository<Account, Integer> {
 
 }
