+++ src/main/java/com/xiia/indicatori/repositories/FixedRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.xiia.indicatori.domain.Fixed;
 
-public interface FixedRepository extends JpaRepository<Fixed, Long> {
+public interface FixedRepository extends JpaRepository<Fixed, Integer> {
 
 	List<Fixed> findAllByAccountId(Integer account);
